pyproject.toml
[project]
name = "dlt-calibration-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lakeshore>=1.9.2",
    "pyserial>=3.5",
]
